backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.2",
    "gtts>=2.5.4",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "openai>=1.108.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.25",
    "uvicorn>=0.35.0",
    "python-dotenv==1.1.1"
]
